unit test improvements
