+++ test/Selenium.WebDriver.Extensions.Core.Tests/Selenium.WebDriver.Extensions.Core.Tests.csproj
@@ -118,7 +118,6 @@
   <ItemGroup>
     <Compile Include="Extensions\StringExtensionsTests.cs" />
     <Compile Include="Extensions\WebDriverExtensionsTests.cs" />
-    <Compile Include="Extensions\WebElementExtensionsTests.cs" />
     <Compile Include="Loaders\QuerySelectorLoaderTests.cs" />
     <Compile Include="Models\ByTests.cs" />
     <Compile Include="Models\WebElementTests.cs" />
